docs

+++ PRUF_TRUFFLE/contracts/PRUF_INTERFACES.sol
@@ -457,16 +457,6 @@ interface AC_TKN_Interface {
         string calldata _tokenURI
     ) external returns (uint256);
 
-    // /*
-    //  * @dev remint AC Token
-    //  * burns old token
-    //  * Sends new token to _recipientAddreass
-    //  */
-    // function reMintACToken(
-    //     address _recipientAddress,
-    //     uint256 tokenId,
-    //     string calldata _tokenURI
-    // ) external returns (uint256);
 
     /**
      * @dev Transfers the ownership of a given token ID to another address.
@@ -608,15 +598,6 @@ interface A_TKN_Interface {
         string calldata _tokenURI
     ) external returns (uint256);
 
-    // /*
-    //  * @dev remint Asset Token
-    //  * must set a new and unuiqe rgtHash
-    //  * burns old token
-    //  * Sends new token to original Caller
-    //  */
-    // function reMintAssetToken(address _recipientAddress, uint256 tokenId)
-    //     external
-    //     returns (uint256);
 
     /*
      * @dev Set new token URI String
@@ -637,7 +618,7 @@ interface A_TKN_Interface {
     /*
      * @dev See if token exists
      */
-    function tokenExists(uint256 tokenId) external view returns (uint8);
+    function tokenExists(uint256 tokenId) external view returns (uint256);
 
     /**
      * @dev Transfers the ownership of a given token ID to another address.
@@ -791,7 +772,7 @@ interface ID_TKN_Interface {
     ) external returns (uint256);
 
     /*
-     * @dev remint Asset Token
+     * @dev remint ID Token
      * must set a new and unuiqe rgtHash
      * burns old token
      * Sends new token to original Caller
@@ -803,7 +784,7 @@ interface ID_TKN_Interface {
     /*
      * @dev See if token exists
      */
-    function tokenExists(uint256 tokenId) external view returns (uint8);
+    function tokenExists(uint256 tokenId) external view returns (uint256);
 
     /**
      * @dev @dev Blocks the transfer of a given token ID to another address
